Abstraction model for JS views.
Polish checks on null and undefined;
Rename application to view;
Update README.md frontend section;
